Moving list test to internal name as it is not public API testing
